Ignore fresh waypoints in handleScroll
In the rare event that a Waypoint was created in the instance
where a scroll value went from negative to 0, that Waypoint would
be triggered because of its null triggerPoint + the combination
of checks that occured during the scroll handler. This change
explicitly ignores waypoints with null trigger points during scroll
checks. Fresh waypoints will be checked anyway during the refresh
that occurs as part of Waypoint instantiation.