Only trigger fresh waypoints that are within the viewable area
It doesn’t make sense to trigger fresh waypoints that are already
outside of the visible area of the scrollable context. The waypoints
may be initialised when the context has already been scrolled somewhat.

This change will only trigger a fresh waypoint if that waypoint is
between the top and bottom of the visible area of the scrollable
context. It triggers both forward and back because there is no concept
of just ‘visible’.